// src_frontend/api-client/models/ExportSegment.ts
/* tslint:disable */
/* eslint-disable */
/**
 *
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 0.0.0
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import type { ExportSegmentColumn } from "./ExportSegmentColumn";
import {
  ExportSegmentColumnFromJSON,
  ExportSegmentColumnToJSON,
} from "./ExportSegmentColumn";

/**
 *
 * @export
 * @interface ExportSegment
 */
export interface ExportSegment {
  /**
   *
   * @type {string}
   * @memberof ExportSegment
   */
  id: string;
  /**
   *
   * @type {string}
   * @memberof ExportSegment
   */
  displayName: string;
  /**
   *
   * @type {string}
   * @memberof ExportSegment
   */
  description: string;
  /**
   *
   * @type {Array<ExportSegmentColumn>}
   * @memberof ExportSegment
   */
  columns: Array<ExportSegmentColumn>;
}

/**
 * Check if a given object implements the ExportSegment interface.
 */
export function instanceOfExportSegment(value: object): value is ExportSegment {
  if (!("id" in value) || value["id"] === undefined) return false;
  if (!("displayName" in value) || value["displayName"] === undefined)
    return false;
  if (!("description" in value) || value["description"] === undefined)
    return false;
  if (!("columns" in value) || value["columns"] === undefined) return false;
  return true;
}

export function ExportSegmentFromJSON(json: any): ExportSegment {
  return ExportSegmentFromJSONTyped(json, false);
}

export function ExportSegmentFromJSONTyped(
  json: any,
  ignoreDiscriminator: boolean,
): ExportSegment {
  if (json == null) {
    return json;
  }
  return {
    id: json["id"],
    displayName: json["display_name"],
    description: json["description"],
    columns: (json["columns"] as Array<any>).map(ExportSegmentColumnFromJSON),
  };
}

export function ExportSegmentToJSON(json: any): ExportSegment {
  return ExportSegmentToJSONTyped(json, false);
}

export function ExportSegmentToJSONTyped(
  value?: ExportSegment | null,
  ignoreDiscriminator: boolean = false,
): any {
  if (value == null) {
    return value;
  }

  return {
    id: value["id"],
    display_name: value["displayName"],
    description: value["description"],
    columns: (value["columns"] as Array<any>).map(ExportSegmentColumnToJSON),
  };
}